+++ spec/ExampleTesterSpec.php
@@ -5,6 +5,7 @@ namespace spec\hanneskod\readmetester;
 
 use hanneskod\readmetester\ExampleTester;
 use hanneskod\readmetester\ListenerInterface;
+use hanneskod\readmetester\Example\ExampleStoreInterface;
 use hanneskod\readmetester\Example\ExampleObj;
 use hanneskod\readmetester\Utils\CodeBlock;
 use hanneskod\readmetester\Runner\OutcomeInterface;
@@ -18,6 +19,7 @@ class ExampleTesterSpec extends ObjectBehavior
 {
     function it_tests_examples_and_fires_events(
         CodeBlock $codeBlock,
+        ExampleStoreInterface $exampleStore,
         ExampleObj $example,
         OutcomeInterface $outcome,
         RunnerInterface $runner,
@@ -36,13 +38,17 @@ class ExampleTesterSpec extends ObjectBehavior
         $listener->onExample($example)->shouldBeCalled();
         $listener->onExpectation($status)->shouldBeCalled();
 
-        $this->beConstructedWith($runner, $evaluator);
+        $this->beConstructedWith($runner, $evaluator, false);
 
         $this->registerListener($listener);
-        $this->testExample($example);
+
+        $exampleStore->getExamples()->willReturn([$example]);
+
+        $this->test($exampleStore);
     }
 
     function it_ignores_examples(
+        ExampleStoreInterface $exampleStore,
         ExampleObj $example,
         RunnerInterface $runner,
         ExpectationEvaluator $evaluator,
@@ -52,9 +58,12 @@ class ExampleTesterSpec extends ObjectBehavior
 
         $listener->onIgnoredExample($example)->shouldBeCalled();
 
-        $this->beConstructedWith($runner, $evaluator);
+        $this->beConstructedWith($runner, $evaluator, false);
 
         $this->registerListener($listener);
-        $this->testExample($example);
+
+        $exampleStore->getExamples()->willReturn([$example]);
+
+        $this->test($exampleStore);
     }
 }